+++ includes/platformer-engine/Engine/Engine.hpp
@@ -6,6 +6,7 @@
 #include "Physics/PhysicsSystem.hpp"
 #include "GameLevel/GameLevel.hpp"
 #include <memory>
+#include "Scene.hpp"
 
 
 namespace platformer_engine {
@@ -33,7 +34,7 @@ namespace platformer_engine {
         void operator=(Engine &&) = delete;
 
         /**
-         * @brief Start the engine and create a window
+         * @brief Initialize the engine
          * @param width The width of the window in px
          * @param height The height of the window in px
          * @param title  Title of the window
@@ -42,6 +43,12 @@ namespace platformer_engine {
          * @platformerengine
          */
         auto Init(int width, int height, const std::string &title, const spic::Color &color) -> bool;
+
+        /**
+         * @brief Start the engine, open window, start timer etc.
+         * @platformerengine
+         */
+        void Start();
 
         /**
          * @brief Render the game on the screen
@@ -66,6 +73,19 @@ namespace platformer_engine {
          */
         void Quit();
 
+        /**
+         * @brief Set the current active Scene
+         * @param scene Scene to make active
+         * @platformerengine
+         */
+        void SetActiveScene(std::unique_ptr<spic::Scene> scene);
+
+        /**
+         * @brief Get the current active Scene
+         * @return std::unique_ptr<spic::Scene>& Current active scene
+         * @platformerengine
+         */
+        auto GetActiveScene() -> std::unique_ptr<spic::Scene>&;
 
     private:
         Engine() = default;