+++ src/Render/Window.cpp
@@ -21,3 +21,11 @@ auto platformer_engine::Window::ListenForEvents() -> std::vector<EventsEnum> {
    return InputFacade::GetInstance().ListenForInput();
 }
 
+void platformer_engine::Window::SetActiveScene(std::unique_ptr<spic::Scene> scene) {
+    _activeScene.swap(scene);
+}
+
+auto platformer_engine::Window::GetActiveScene() -> std::unique_ptr<spic::Scene> & {
+    return _activeScene;
+}
+